bugfixes: - base64 fixed to accept unpadded input - revert should not deleted empty directories that was added - base64 input string should always be trimmed

+++ javasvn/src/org/tmatesoft/svn/core/internal/wc/SVNDirectory.java
@@ -1191,7 +1191,9 @@ public class SVNDirectory {
         dir.getEntries().save(true);
 
         SVNFileUtil.deleteAll(new File(dir.getRoot(), dir.getAdminDirectory().getName()), dir.getWCAccess());
-        dir.getRoot().delete();
+        if (deleteWorkingFiles) {
+            dir.getRoot().delete();
+        }
     }
 
     public void commit(String target, SVNCommitInfo info, Map wcPropChanges,
